Added the order status to check if the user's balance can be updated

+++ src/main/java/com/tlc/group/seven/orderprocessingservice/portfolio/model/Portfolio.java
@@ -16,7 +16,7 @@ public class Portfolio {
     private Long iD;
     @NotNull
     private String ticker;
-    private Double quantity;
+    private Integer quantity;
 
     @JsonIgnore
     @ManyToOne(fetch = FetchType.LAZY)
